--- a/.gitignore
+++ b/.gitignore
@@ -1,2 +1,3 @@
 /kubectl-opslevel
 src/dist
+.idea