--- a/src/go.mod
+++ b/src/go.mod
@@ -13,6 +13,7 @@ require (
 	github.com/rs/zerolog v1.21.0
 	github.com/spf13/cobra v1.2.1
 	github.com/spf13/viper v1.8.1
+	go.uber.org/automaxprocs v1.4.0
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
 	k8s.io/apimachinery v0.20.2
 	k8s.io/client-go v0.20.2